+++ cztctl/rpc/sdk/vars.go
@@ -4,7 +4,7 @@ const (
 	// DefaultRepoUser is the default repo auth username for push.
 	DefaultRepoUser = "cztctl-bot"
 	// DefaultRepoToken is the default repo auth credential (password/token/key) for push.
-	DefaultRepoToken = "glpat-Cvi30LCsPVteDGdPWxmER286MQp1OjUH.01.0w0iuu5yw"
+	DefaultRepoToken = "glpat-rinPIxpXXi0v7pAUKAKQTG86MQp1OjEH.01.0w13qos5x"
 	// DefaultRemote is the default remote template repository (empty means use goctl default).
 	DefaultRemote = ""
 	// DefaultRepoBranch is the default git branch name for SDK repo.